Battery recharge and patches
Addition of a battery recharge system which the roomba will gradually run out of battery and will recharge at a specific spot. Other than that it was added a system of which creates a lot of panels working as a grid. When the roomba goes trough one, it will clean it. We can make it random later or smth.