Auto Movement and ui elements
Now the thing goes around automatic. It'll start moving south and will know to back up if it encounters a wall or a cliff. If the not a single dirty patch is found in the path it chose, it will change direction. There are known issues regarding the cliff recognition (Sometimes when going south it will not recognize in time the cliff). There was also the addition of ui elements to know the batery number and a restart button.